Non stop mode for strength endurance exercises
For endurance exercises, detection of the end of the exercise breaks the
measures, stopping the analysis.

Created a new command to tell the backend that it should not detected
the end of the exercise.
This allows to run endurance exercises where the strength is released
completly during short intervals.

Removed some old references to body weight in the backend

Fix open browser